refactor: colocate container, font, head, heading, hr, html, img in react-email
Point each @react-email/* package build at shared source; add vitest configs.

Made-with: Cursor
