+++ packages/font/vitest.config.ts
@@ -1,0 +1,9 @@
+import { defineConfig } from 'vitest/config';
+
+export default defineConfig({
+  test: {
+    globals: true,
+    environment: 'happy-dom',
+    include: ['../react-email/src/components/font/**/*.spec.{ts,tsx}'],
+  },
+});
