+++ packages/react-email/src/components/container/index.ts
@@ -1,0 +1,1 @@
+export * from './container.js';
