+++ packages/react-email/src/components/font/index.ts
@@ -1,0 +1,1 @@
+export * from './font.js';
