+++ packages/react-email/src/components/head/index.ts
@@ -1,0 +1,1 @@
+export * from './head.js';
