+++ packages/react-email/src/components/heading/heading.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
-import type { As } from './utils/as';
-import type { Margin } from './utils/spaces';
-import { withMargin } from './utils/spaces';
+import type { As } from './utils/as.js';
+import type { Margin } from './utils/spaces.js';
+import { withMargin } from './utils/spaces.js';
 
 export type HeadingAs = As<'h1', 'h2', 'h3', 'h4', 'h5', 'h6'>;
 export type HeadingProps = HeadingAs & Margin;
