+++ packages/react-email/src/components/heading/index.ts
@@ -1,0 +1,1 @@
+export * from './heading.js';
