+++ packages/react-email/src/components/hr/index.ts
@@ -1,0 +1,1 @@
+export * from './hr.js';
