+++ packages/react-email/src/components/html/index.ts
@@ -1,0 +1,1 @@
+export * from './html.js';
