+++ packages/react-email/src/components/img/index.ts
@@ -1,0 +1,1 @@
+export * from './img.js';
