+++ packages/react-email/src/components/index.ts
@@ -3,3 +3,10 @@ export * from './button/index.js';
 export * from './code-block/index.js';
 export * from './code-inline/index.js';
 export * from './column/index.js';
+export * from './container/index.js';
+export * from './font/index.js';
+export * from './head/index.js';
+export * from './heading/index.js';
+export * from './hr/index.js';
+export * from './html/index.js';
+export * from './img/index.js';
